+++ tests/Bridge/Symfony/Bundle/DependencyInjection/ApiPlatformExtensionTest.php
@@ -1227,6 +1227,7 @@ class ApiPlatformExtensionTest extends TestCase
             'api_platform.graphql.graphql_playground.enabled' => true,
             'api_platform.resource_class_directories' => Argument::type('array'),
             'api_platform.validator.serialize_payload_fields' => [],
+            'api_platform.validator.query_parameter_validation' => true,
             'api_platform.elasticsearch.enabled' => false,
             'api_platform.asset_package' => null,
             'api_platform.defaults' => ['attributes' => []],